+++ app/core/list/index.ts
@@ -1,2 +1,3 @@
 export * from "./list-base.component";
 export * from "./list-selection";
+export * from "./selectable-list";
